feat(audience): Devide audience UI to (components, page, ...)
features/audience/
├── components/
│   ├── MultipleChoiceInput.tsx
│   ├── OpenEndedInput.tsx
│   └── AudienceOpinionPage.tsx
├── hooks/
│   ├── useQuestions.ts
│   └── useTimer.ts
├── pages/
│   └── index.tsx
├── services/
│   └── questionService.ts
├── types/
│   └── index.ts
├── utils/
│   └── shuffle.ts
├── constants.ts
├── routes.ts
└── index.ts
