+++ src/features/audience/constants.ts
@@ -1,0 +1,6 @@
+export const QUESTION_TYPES = {
+  MULTIPLE_CHOICE: 'multiple-choice' as const,
+  OPEN_ENDED: 'open-ended' as const,
+};
+
+export const DEFAULT_TIME = 60;
