+++ src/features/audience/index.ts
@@ -1,0 +1,4 @@
+export { default as AudienceOpinionPage } from '../audience/pages/AudienceOpinionPage';
+export * from './types';
+export * from './constants';
+export * from './routes';
