+++ src/features/audience/routes.ts
@@ -1,0 +1,3 @@
+export const AUDIENCE_ROUTES = {
+  OPINION_PAGE: '/audience/opinion',
+};
